Rename InfoVAE -> VAE and fix docs
