+++ bayesflow/experimental/autoencoder/variational_autoencoder.py
@@ -1,24 +1,52 @@
 import keras
 
+from bayesflow.metrics.functional import maximum_mean_discrepancy
 from bayesflow.types import Tensor
-from bayesflow.utils import filter_kwargs, weighted_mean
+from bayesflow.utils import resolve_seed, non_batch_axis, weighted_mean
 from bayesflow.utils.serialization import serializable, serialize
-
 from .autoencoder import AutoEncoder
 
 
 @serializable("bayesflow.experimental")
 class VariationalAutoEncoder(AutoEncoder):
-    r"""Variational autoencoder, which compresses data into a low-dimensional latent representation.
-    Applies a variational assumption; i.e., the forward pass is stochastic under the assumption that the latent vector
+    """Information-Maximizing Variational Autoencoder according to [1].
 
-    :math:`z = E(x) \sim \mathcal{N}(\mu(x), \sigma(x)^2)`
+    The loss is computed as
 
-    where :math:`\mu` and :math:`\sigma^2` are the variational parameters inferred from the input :math:`x`.
+        loss = reconstruction_loss
+             + w_kl  * KL[q(z | x) || p(z)]
+             + w_mmd * MMD[q(z), p(z)]
+    with
+        w_kl  = 1 - alpha
+        w_mmd = alpha + beta - 1
 
-    The inverse pass is deterministic:
+    Useful settings are:
 
-    :math:`x = D(z)`
+        Vanilla VAE (default):  w_kl=1,    w_mmd=0    -> alpha=0,        beta=1
+        beta-VAE:               w_kl=beta, w_mmd=0    -> alpha=1-beta,   beta=beta
+        MMD/InfoVAE:            w_kl=0,    w_mmd=lam  -> alpha=1,        beta=lam
+        Mixed objective:        w_kl=a,    w_mmd=b    -> alpha=1-a,      beta=a+b
+
+    [1] Zhao, S., Song, J., & Ermon, S. (2019). InfoVAE: Balancing learning and
+    inference in variational autoencoders. In Proceedings of the AAAI Conference on
+    Artificial Intelligence (Vol. 33, No. 01, pp. 5885-5892).
+
+    Parameters
+    ----------
+    latent_dim
+        Dimensionality of the latent variable.
+    encoder_network
+        Network mapping inputs to an encoder representation.
+    decoder_network
+        Network mapping latent samples to a decoder representation.
+    alpha
+        InfoVAE information parameter. Controls the weight of the conditional
+        encoder KL through ``1 - alpha``.
+    beta
+        InfoVAE aggregate-posterior matching parameter. Together with alpha,
+        controls the MMD weight through ``alpha + beta - 1``. In [1], this parameter is named lambda.
+    mmd_kwargs
+        Optional keyword arguments forwarded to ``maximum_mean_discrepancy``.
     """
 
     def __init__(
@@ -26,57 +54,138 @@ class VariationalAutoEncoder(AutoEncoder):
         latent_dim: int,
         encoder_network: keras.Layer,
         decoder_network: keras.Layer,
-        kl_weight: float = 1e-6,
-        recon_weight: float = 1.0,
+        alpha: float = 0.0,
+        beta: float = 1.0,
+        mmd_kwargs: dict | None = None,
         **kwargs,
     ):
-        super().__init__(latent_dim, encoder_network, decoder_network, **kwargs)
+        super().__init__(
+            latent_dim=latent_dim,
+            encoder_network=encoder_network,
+            decoder_network=decoder_network,
+            **kwargs,
+        )
 
-        # adjust the number of output units to account for the split into mean and variance
         self.encoder_projector.units = 2 * latent_dim
+        self.alpha = alpha
+        self.beta = beta
+        self.mmd_kwargs = mmd_kwargs or {}
 
-        self.kl_weight = kl_weight
-        self.recon_weight = recon_weight
         self.seed_generator = keras.random.SeedGenerator()
+
+    @property
+    def kl_weight(self) -> float:
+        return 1.0 - self.alpha
+
+    @property
+    def mmd_weight(self) -> float:
+        return self.alpha + self.beta - 1.0
 
     def get_config(self):
         base_config = super().get_config()
-        config = {
-            "kl_weight": self.kl_weight,
-            "recon_weight": self.recon_weight,
-        }
+        config = {"alpha": self.alpha, "lambd": self.beta, "mmd_kwargs": self.mmd_kwargs}
         return base_config | serialize(config)
 
     def _encode(
-        self, x: Tensor, training: bool = False, seed: int | keras.random.SeedGenerator | None = None, **kwargs
+        self,
+        x: Tensor,
+        training: bool = False,
+        seed: int | keras.random.SeedGenerator | None = None,
+        **kwargs,
     ):
-        if seed is None:
-            seed = self.seed_generator
+        seed = resolve_seed(seed)
 
-        y = self.encoder_network(x, training=training, **filter_kwargs(kwargs, self.encoder_network.call))
-        z = self.encoder_projector(y, training=training, **filter_kwargs(kwargs, self.encoder_projector.call))
+        z = super()._forward(x, training=training, **kwargs)
         mean, log_var = keras.ops.split(z, 2, axis=-1)
-        epsilon = keras.random.normal(keras.ops.shape(mean), seed=seed, dtype=mean.dtype)
-        sample = mean + keras.ops.exp(log_var / 2) * epsilon
+
+        epsilon = keras.random.normal(
+            shape=keras.ops.shape(mean),
+            seed=seed,
+            dtype=mean.dtype,
+        )
+
+        sample = mean + keras.ops.exp(0.5 * log_var) * epsilon
+
         return z, mean, log_var, epsilon, sample
 
-    def _forward(self, x: Tensor, training: bool = False, **kwargs):
-        *_, sample = self._encode(x, training=training, **kwargs)
+    def _forward(
+        self,
+        x: Tensor,
+        training: bool = False,
+        seed: int | keras.random.SeedGenerator | None = None,
+        **kwargs,
+    ):
+        *_, sample = self._encode(
+            x,
+            training=training,
+            seed=seed,
+            **kwargs,
+        )
         return sample
 
+    def _conditional_kl(self, mean: Tensor, log_var: Tensor) -> Tensor:
+        """Per-example KL[q(z | x) || p(z)] for diagonal Gaussian q and standard normal p."""
+
+        return 0.5 * keras.ops.sum(
+            keras.ops.square(mean) + keras.ops.exp(log_var) - 1.0 - log_var,
+            axis=non_batch_axis(mean),
+        )
+
+    def _marginal_mmd(self, sample: Tensor, seed: int | keras.random.SeedGenerator | None = None) -> Tensor:
+        """MMD[q(z), p(z)] using samples from the aggregate posterior and prior."""
+
+        seed = resolve_seed(seed)
+        targets = keras.random.normal(
+            shape=keras.ops.shape(sample),
+            seed=seed,
+            dtype=sample.dtype,
+        )
+
+        return maximum_mean_discrepancy(sample, targets, **self.mmd_kwargs)
+
+    def _reconstruction_loss(self, x: Tensor, reconstruction: Tensor) -> Tensor:
+        """Per-example mean squared reconstruction error."""
+
+        return keras.ops.mean(
+            keras.ops.square(x - reconstruction),
+            axis=non_batch_axis(x),
+        )
+
     def compute_metrics(
-        self, x: Tensor, sample_weight: Tensor = None, stage: str = "training", **kwargs
+        self,
+        x: Tensor,
+        sample_weight: Tensor = None,
+        stage: str = "training",
+        seed: int | keras.random.SeedGenerator | None = None,
+        **kwargs,
     ) -> dict[str, Tensor]:
         training = stage == "training"
+        seed = resolve_seed(seed)
 
-        z, mean, log_var, epsilon, sample = self._encode(x, training=training, **kwargs)
-        reconstruction = self(sample, training=training, inverse=True, **kwargs)
-
-        non_batch_axes = list(range(1, keras.ops.ndim(mean)))
-        recon_loss = self.recon_weight * keras.ops.mean(keras.ops.square(x - reconstruction), axis=non_batch_axes)
-        kl_loss = self.kl_weight * keras.ops.sum(
-            -0.5 * (1.0 + log_var - keras.ops.square(mean) - keras.ops.exp(log_var)), axis=non_batch_axes
+        _, mean, log_var, _, sample = self._encode(
+            x,
+            training=training,
+            seed=seed,
+            **kwargs,
         )
-        loss = kl_loss + recon_loss
-        loss = weighted_mean(loss, sample_weight)
-        return {"loss": loss, "recon_loss": recon_loss, "kl_loss": kl_loss, "z": sample}
+
+        reconstruction = self(
+            sample,
+            training=training,
+            inverse=True,
+            **kwargs,
+        )
+
+        recon_loss = weighted_mean(self._reconstruction_loss(x, reconstruction), sample_weight)
+        kl_loss = keras.ops.mean(self._conditional_kl(mean, log_var))
+
+        loss = recon_loss + self.kl_weight * kl_loss
+
+        if self.mmd_weight != 0.0:
+            mmd_targets = keras.random.normal(shape=keras.ops.shape(sample), seed=seed, dtype=sample.dtype)
+            marginal_mmd = maximum_mean_discrepancy(sample, mmd_targets, **self.mmd_kwargs)
+            loss = loss + self.mmd_weight * marginal_mmd
+        else:
+            marginal_mmd = keras.ops.zeros((), dtype=sample.dtype)
+
+        return {"loss": loss, "recon_loss": recon_loss, "kl_loss": kl_loss, "mmd_loss": marginal_mmd, "z": sample}
